refactor: Extract to service
